+++ tests/framework/db/mssql/conditions/LikeConditionBuilderTest.php
@@ -38,7 +38,7 @@ final class LikeConditionBuilderTest extends BaseDatabase
 
         $db = $this->getConnection(false, false);
 
-        [$sql, $params] = $db->getQueryBuilder()->build($query);
+        [$sql, $params] = $db->queryBuilder->build($query);
 
         self::assertSame(
             'SELECT *' . ($expected === '' ? '' : ' WHERE ' . $this->replaceQuotes($expected)),