+++ tests/framework/db/oci/conditions/LikeConditionBuilderTest.php
@@ -57,7 +57,7 @@ final class LikeConditionBuilderTest extends BaseDatabase
 
         $query = (new Query())->where($condition);
 
-        [$sql, $params] = $db->getQueryBuilder()->build($query);
+        [$sql, $params] = $db->queryBuilder->build($query);
 
         self::assertSame(
             'SELECT *' . ($expected === '' ? '' : ' WHERE ' . $this->replaceQuotes($expected)),